working keep/drop high/low
Implemented selectors as SetSelector, subclass of ExprResult that does not have a value.
ConditionalSelector, matching using lambdas that take set items as input, will be inherited from for comparisons selectors.

Keep/Drop operators and High/Low selectors are functioning with the new set-based dice.

Formatting needs to be redone for this new representation.